moved constants into the rapidata project to fix import issues
