+++ src/rapidata/constants.py
